+++ src/tests/mod.rs
@@ -6,8 +6,8 @@ use crate::{
     extract::{self, Path},
     handler::Handler,
     response::IntoResponse,
-    routing::{any, delete, get, on, patch, post, MethodFilter},
-    service, Json, Router,
+    routing::{any, delete, get, on, patch, post, service_method_router as service, MethodFilter},
+    Json, Router,
 };
 use bytes::Bytes;
 use http::{